requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
sqlmodel==0.0.22
httpx==0.27.0
stripe==10.12.0
markdown==3.6